Extract test component to separate file
