+++ webApp/src/jsMain/typescript/io/rwc/streamwise/app/app.ts
@@ -2,19 +2,7 @@ import {Component} from '@angular/core';
 import {RouterOutlet} from '@angular/router';
 
 import * as streamwise from '@streamwise';
-import {BaseChartDirective} from "ng2-charts";
-
-@Component({
-  templateUrl: './test.component.html',
-  standalone: true,
-  imports: [BaseChartDirective],
-  selector: 'test-component',
-})
-class TestComponent extends streamwise.TestComponent {
-  constructor() {
-    super()
-  }
-}
+import {TestComponent} from "./components/test.component";
 
 @Component({
   selector: 'app-root',
